fix console script execution
