+++ cats/cats.py
@@ -1,13 +1,18 @@
 import click
+import aiohttp
 import asyncio
 import re
 import json
 
-from typing import Tuple, Iterable, Union, List
+from typing import Optional, Tuple, Iterable, Union, List
 from blspy import G2Element, AugSchemeMPL
 
 from chia.cmds.wallet_funcs import get_wallet
 from chia.rpc.wallet_rpc_client import WalletRpcClient
+from chia.rpc.wallet_rpc_client import WalletRpcClient
+from chia.util.default_root import DEFAULT_ROOT_PATH
+from chia.util.config import load_config
+from chia.util.ints import uint16
 from chia.util.byte_types import hexstr_to_bytes
 from chia.types.blockchain_format.program import Program
 from clvm_tools.clvmc import compile_clvm_text
@@ -20,8 +25,39 @@ from chia.wallet.cat_wallet.cat_utils import (
     unsigned_spend_bundle_for_spendable_cats,
 )
 from chia.util.bech32m import decode_puzzle_hash
-from secure_the_bag import read_secure_the_bag_targets, secure_the_bag
-from utils import get_client, get_signed_tx
+from cats.secure_the_bag import read_secure_the_bag_targets, secure_the_bag
+
+
+# Loading the client requires the standard chia root directory configuration that all of the chia commands rely on
+async def get_client() -> Optional[WalletRpcClient]:
+    try:
+        config = load_config(DEFAULT_ROOT_PATH, "config.yaml")
+        self_hostname = config["self_hostname"]
+        full_node_rpc_port = config["wallet"]["rpc_port"]
+        full_node_client = await WalletRpcClient.create(
+            self_hostname, uint16(full_node_rpc_port), DEFAULT_ROOT_PATH, config
+        )
+        return full_node_client
+    except Exception as e:
+        if isinstance(e, aiohttp.ClientConnectorError):
+            print(
+                f"Connection error. Check if full node is running at {full_node_rpc_port}"
+            )
+        else:
+            print(f"Exception from 'harvester' {e}")
+        return None
+
+
+async def get_signed_tx(fingerprint, ph, amt, fee):
+    try:
+        wallet_client: WalletRpcClient = await get_client()
+        wallet_client_f, _ = await get_wallet(wallet_client, fingerprint)
+        return await wallet_client_f.create_signed_transaction(
+            [{"puzzle_hash": ph, "amount": amt}], fee=fee
+        )
+    finally:
+        wallet_client.close()
+        await wallet_client.await_closed()
 
 
 async def push_tx(fingerprint, bundle):
